+++ sdk/samples/iot/sample_pnp_component_mqtt.c
@@ -15,7 +15,7 @@ static char request_id_buf[10];
 az_span get_request_id(void)
 {
   az_span remainder;
-  az_span out_span = az_span_init((uint8_t*)request_id_buf, sizeof(request_id_buf));
+  az_span out_span = az_span_create((uint8_t*)request_id_buf, sizeof(request_id_buf));
 
   // Note that if left to run for a long time, this will overflow and reset back to 0.
   az_result result = az_span_u32toa(out_span, request_id_int++, &remainder);