+++ sdk/docs/core/CHANGELOG.md
@@ -4,6 +4,11 @@
 ## 1.0.0-preview.4 (Unreleased)
 
 - API breaking changes:
+  - `az_span.h`:
+    - `az_span_init()` is renamed to `az_span_create()`.
+    - `az_span_from_str()` is renamed to `az_span_create_from_str()`.
+    - `az_allocator_context` is renamed to `az_span_allocator_context`.
+    - Removed `az_pair_from_str()`.
   - `az_context`:
     - `key` and `value` are `const`.
     - `az_context_with_expiration()` is renamed to `az_context_create_with_expiration()`.
@@ -14,7 +19,7 @@
     - `status_codes` now should be terminated by `AZ_HTTP_STATUS_CODE_END_OF_LIST`.
     - `max_retries` is now `int32_t` instead of `int16_t`.
   - `az_config.h`:
-    -  `AZ_HTTP_REQUEST_URL_BUF_SIZE` renamed to `AZ_HTTP_REQUEST_URL_BUFFER_SIZE`.
+    - `AZ_HTTP_REQUEST_URL_BUF_SIZE` renamed to `AZ_HTTP_REQUEST_URL_BUFFER_SIZE`.
     - `AZ_HTTP_REQUEST_BODY_BUF_SIZE` renamed to `AZ_HTTP_REQUEST_BODY_BUFFER_SIZE`.
     - `AZ_LOG_MSG_BUF_SIZE` renamed to `AZ_LOG_MESSAGE_BUFFER_SIZE`.